zacatek velkeho uklidu
