+++ app/templates/Sprava/nastroje.latte
@@ -4,15 +4,15 @@
     <h3>Nástroje pro správu</h3>
     
     {if $canCreateArea}
-        <a href="{link Sprava:novaoblast id=>null}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-pencil icon-pencil"></i> Vytvořit oblast</a>
-        <a href="{link Sprava:noveap id=>null}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-pencil icon-pencil"></i> Vytvořit AP</a>
+        <a href="{link SpravaOblasti:novaoblast id=>null}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-pencil icon-pencil"></i> Vytvořit oblast</a>
+        <a href="{link SpravaOblasti:noveap id=>null}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-pencil icon-pencil"></i> Vytvořit AP</a>
         <br><br>
     {/if}
     {if $canApproveCC}
         <a href="{link Sprava:slucovani id=>null}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-pencil icon-pencil"></i> Sloučit uživatele</a>
         <a href="{link Sprava:slouceni id=>null}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-pencil icon-pencil"></i> Sloučení uživatelé</a>
         <br><br>
-        <a href="{link Sprava:presifrovani id=>null}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-pencil icon-pencil"></i> Přešifrování hesel</a>
+        <a href="{link SpravaSifrovani:presifrovani id=>null}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-pencil icon-pencil"></i> Přešifrování hesel</a>
         <br><br>
         <a href="{link Sprava:schvalovanicc id=>null}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-pencil icon-pencil"></i> Schvalování čestného členství</a>
